Move the measurements class to utility.hpp
